Fixed some GUI things
Fixed the toString in Rider so that it would use ints instead of
doubles.

I also collected data, however, we may want to recollect the data.

+++ Rider.h
@@ -41,7 +41,7 @@ public:
 		double d = pass.tot_wait_chairlifts() + pass.tot_wait_trailheads();
 		std::string s;
 		s.append("Skill Level: " + std::to_string((int)skill_lvl) + "\nEndurance: " + std::to_string((int)endurance)  + "\nTotal Wait Time: "
-			+ std::to_string((int)d) + "\nAverage Trailhead Wait Time: " + std::to_string((double)pass.AvgTrailWait) + "\nTotal Trailhead Wait Time: " + std::to_string((double)pass.TotTrailWait) + "\nAverage Chairlift Wait Time: " + std::to_string((double)pass.AvgChairWait) + "\nTotal Chairlift Wait Time: " + std::to_string((double)pass.TotChairWait));
+			+ std::to_string((int)d) + "\nAverage Trailhead Wait Time: " + std::to_string((int) pass.AvgTrailWait) + "\nTotal Trailhead Wait Time: " + std::to_string((int)pass.TotTrailWait) + "\nAverage Chairlift Wait Time: " + std::to_string((int)pass.AvgChairWait) + "\nTotal Chairlift Wait Time: " + std::to_string((int)pass.TotChairWait));
 		return s;
 	}
 };
